pyproject.toml
[project]
name = "simple-a2a-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.0",
    "asyncclick>=8.1.8",
    "faiss-cpu>=1.11.0.post1",
    "google-adk>=1.9.0",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.9",
    "langchain-huggingface>=0.3.1",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-unstructured>=0.1.5",
    "langgraph>=0.6.3",
    "mcp[cli]>=1.12.3",
    "sentence-transformers>=5.0.0",
    "unstructured[pdf]>=0.18.11",
]
